+++ frontend/src/pages/ManagerLogin/units/LoginForm.tsx
@@ -1,77 +1,29 @@
-import { AxiosError, AxiosResponse } from 'axios';
-import { FormEventHandler, useState } from 'react';
-import { useMutation } from 'react-query';
-import { useHistory } from 'react-router';
-import { useSetRecoilState } from 'recoil';
-import { postLogin } from 'api/login';
 import Button from 'components/Button/Button';
 import Input from 'components/Input/Input';
 import MANAGER from 'constants/manager';
-import MESSAGE from 'constants/message';
-import PATH from 'constants/path';
-import { LOCAL_STORAGE_KEY } from 'constants/storage';
 import useInputs from 'hooks/useInputs';
-import accessTokenState from 'state/accessTokenState';
-import { ErrorResponse, LoginSuccess } from 'types/response';
-import { setLocalStorageItem } from 'utils/localStorage';
-import * as Styled from './LoginForm.styles';
+import { ErrorMessage, LoginParams } from '../ManagerLogin';
 
-interface ErrorMessage {
-  email?: string;
-  password?: string;
-}
+import * as Styled from './LoginForm.styles';
 
 interface Form {
   email: string;
   password: string;
 }
 
-const LoginForm = (): JSX.Element => {
-  const history = useHistory();
-  const setAccessToken = useSetRecoilState(accessTokenState);
+interface Props {
+  errorMessage: ErrorMessage;
+  onSubmit: (event: React.FormEvent<HTMLFormElement>, { email, password }: LoginParams) => void;
+}
 
+const LoginForm = ({ errorMessage, onSubmit }: Props): JSX.Element => {
   const [{ email, password }, onChangeForm] = useInputs<Form>({
     email: '',
     password: '',
   });
 
-  const [errorMessage, setErrorMessage] = useState<ErrorMessage>({
-    email: '',
-    password: '',
-  });
-
-  const login = useMutation(postLogin, {
-    onSuccess: (response: AxiosResponse<LoginSuccess>) => {
-      const { accessToken } = response.data;
-
-      setLocalStorageItem({ key: LOCAL_STORAGE_KEY.ACCESS_TOKEN, item: accessToken });
-      setAccessToken(accessToken);
-
-      history.push(PATH.MANAGER_MAIN);
-    },
-    onError: (error: AxiosError<ErrorResponse>) => {
-      const field = error.response?.data.field;
-      const message = error.response?.data.message;
-
-      if (field && message) {
-        setErrorMessage({ [field]: message });
-        return;
-      }
-
-      setErrorMessage({ password: message ?? MESSAGE.LOGIN.UNEXPECTED_ERROR });
-    },
-  });
-
-  const handleSubmit: FormEventHandler<HTMLFormElement> = (event) => {
-    event.preventDefault();
-
-    if (!(email && password)) return;
-
-    login.mutate({ email, password });
-  };
-
   return (
-    <Styled.Form onSubmit={handleSubmit}>
+    <Styled.Form onSubmit={(event) => onSubmit(event, { email, password })}>
       <Input
         type="email"
         label="이메일"
